Safari-611.1.10/LayoutTests/imported/w3c/web-platform-tests/html/canvas/offscreen/pixel-manipulation/2d.imageData.put.unaffected.worker-expected.txt


PASS putImageData() is not affected by context state
